RSFNet_RGBTSeg/requirements.txt
absl-py==0.11.0
addict==2.4.0
apex==0.1
appdirs==1.4.4
autopep8==1.5.5
blessed==1.20.0
cachetools==4.2.1
certifi==2021.5.30
chardet==4.0.0
cityscapesScripts==2.2.0
coloredlogs==15.0.1
contextlib2==21.6.0
cycler==0.10.0
ffmpeg-python==0.2.0
future==0.18.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gpustat==1.0.0
grpcio==1.36.0
humanfriendly==9.2
idna==2.10
joblib==1.1.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
mmcv==1.3.17
mmcv-full==1.3.11
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
nvidia-ml-py==11.495.46
oauthlib==3.1.0
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-contrib-python==4.7.0.72
packaging==21.0
pandas==1.3.3
Pillow==7.1.2
protobuf==3.15.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
timm==0.5.4
toml==0.10.2
torch==1.8.0
torch-dct==0.1.5
torch-tb-profiler==0.2.1
torchsummary==1.5.1
torchvision==0.9.0
tqdm==4.54.1
typing==3.7.4.3
typing-extensions @ file:///Users/ktietz/demo/mc3/conda-bld/typing_extensions_1629887998905/work
urllib3==1.26.3
wcwidth==0.2.6
Werkzeug==1.0.1
yapf==0.31.0
